List of all functions

Here an overview of all functions:

HPSAnalysis.CreateStartConfiguration — Method

CreateStartConfiguration(SimulationName::String, Path::String, BoxSize::Vector{R}, Proteins::Vector{String}, Sequences::Vector{String} ; Axis=`y`, Regenerate=true,SimulationType="Calvados2",ProteinToDomain=Dict(),ProteinToCif=Dict())

Creates the file structure and initialises particle positions for the given parameters.

Arguments:

SimulationName::String: The name of the simulation.

Path::String: The base directory where simulation data will be stored.

BoxSize::Vector{R}: A vector defining the box dimensions (x, y, z).

Proteins::Vector{String}: List of protein names used in the simulation.

Sequences::Vector{String}: List of amino acid sequences corresponding to the proteins.

Optional Arguments:

Axis::String: The axis along which the system is unfolded.

Regenerate::Bool: If true, regenerates initial positions using the Polyply package.

SimulationType::String: Type of Simulation (default is Calvados2).

ProteinToDomain::Dict: Dictionary of domains for the proteins.

ProteinToCif::Dict: Dictionary of the AlphaFold data files for each protein.

Returns:

A tuple (pos, Data) containing the initial positions and the simulation data structure.

HPSAnalysis.SaveAllData — Method

SaveAllData(Sim::SimData{R,I}) where {R<:Real, I<:Integer}

Save of the Field information as a jld2 file.

Arguments:

Sim::SimData{R,I}: A simulation data structure containing the Simulation information.

Creat:

Save File.

HPSAnalysis.computeAvgBondAngles — Method

computeAvgBondAngles(Sim::SimData{T, I}; NBins = 180) where {T<:Real, I<:Integer}

Computes the average bond angles.

Computes the bond angles, the cosine thereof for each triplet of beads i, i+1, i+2 along the backbone of the proteins for each frame. The average bond angle and the average cosine of the bond angle are computed over all steps for each triplet independly. The histogram averages also over all triplets. The persistence length is computed if the average of the cosine of the bond angle is large than zero and otherwise set to 0.

Arguments:

Sim::SimData{R,I}: A simulation data structure containing the Simulation information.

NBins::Int: Number of bins for the angle histogram (default: 180).

Creates:

Sim.AvgBondAngles: Stores the averaged bond angles.

Sim.AvgCosBondAngles: Stores the averaged cosine of bond angles.

Sim.LocalPersistenceLength: Computes the persistence length from bond angles.

Sim.BondAngleHist: Stores the histogram of bond angles.

HPSAnalysis.computeCOM! — Method

computeCOM!(Sim::HPSAnalysis.SimData{R,I}) where {R<:Real, I<:Integer}

Computes the center of mass (COM) for each polymer chain.

Arguments:

Sim::SimData{R,I}: A simulation data structure containing the Simulation information.

Updates:

Sim.COM: COM of each chain.

Sim.COM_uw: Unwrapped COM.

HPSAnalysis.computeCOMOfLargestCluster — Method

computeCOMOfLargestCluster(Sim::SimData{T,I} ) where {I<:Integer, T<:Real}

Computes COM of the largest cluster along the axis Sim.SlabAxis in the unwrapped positions.

For the algorithm to work, it is important that the dense phase of the first frame is not crossing the periodic boundaries. The algorithm take dense phase center of the previous step to pre center the data to about issues at the boundaries. Therefore it works only when the displacement along Sim.SlabAxis between consecutive frames is smaller than half of the box length.

Arguments:

Sim::SimData{R,I}: A simulation data structure containing the Simulation information.

Returns:

Coms::Vector{Float}: COM of largest cluster along the Sim.SlabAxis for each step.

HPSAnalysis.computeCOMsOfCluster — Method

computeCOMsOfCluster(Sim::SimData{T,I} ) where {I<:Integer, T<:Real}

Computes COM for each cluster in each frame.

Arguments:

Sim::SimData{R,I}: A simulation data structure containing the Simulation information.

Returns:

Coms::Matrix{Float}: Matrix of COM for each cluster and time step.

HPSAnalysis.computeClustersByBeadDistance — Method

computeClustersByBeadDistance(Sim::HPSAnalysis.SimData{T,Int}; Cutoff=5.0) where{T<:Real, Int<:Integer}

Computes clusters based on minimal distances between amino acids of protein and cutoff.

Computes graph network for each frame in EquilibrationTime:RGMeasureStep:NSteps by naivly comparing all pairs of proteins, computing the minimal distance between any two amino acids of those proteins and adding an edge in the network if the distance is below the given cutoff. Returns the concatenation of groups of IDs belonging to weekly connected graph components of each frame.

Arguments:

Sim::SimData{R,I}: A simulation data structure containing the Simulation information.

Cutoff::Float: Distance criterion for a Cluster.

Returns:

Cluster::Vector{Vector{Vector{Int}}}: List of Clusters of the chains for each step.

HPSAnalysis.computeClustersByChainCOM — Method

computeClustersByCOM(Sim::HPSAnalysis.SimData{T,Int}; Cutoff=50.0) where{T<:Real, Int<:Integer}

Computes clusters based on distances between protein COMs and cutoff.

Computes graph network for each frame in EquilibrationTime:RGMeasureStep:NSteps by naivly computing minimal distances between all proteins center of masses based on the given cutoff. Returns the concatenation of groups of IDs belonging to weekly connected graph components of all frames.

Arguments:

Sim::SimData{R,I}: A simulation data structure containing the Simulation information.

Cutoff::Float: Distance criterion for a Cluster.

Returns:

Cluster::Vector{Vector{Vector{Int}}}: List of Clusters of the chains.

HPSAnalysis.computeDensityHistogram — Method

computeDensityHistogram(Sim::SimData{R,I}, DivLength=I(10)) where {R<:Real, I<:Integer}

Computes a logarithmic histogram of densities of the subcubes of the simulation box.

The simulation box is divided into DivLength^4 subboxes where each dimension is divided into DivLength many subsections and the axes according to Sim.SlabAxis is divided into DivLength^2 many subsections.

Subboxes that are empty or those that contain the boundary of the simulation box are not considered for the histogram.

Arguments:

Sim::SimData{R,I}: A simulation data structure containing the Simulation information.

DivLength=I(10)): Division factor determining the length of each subbox.

Creat:

Sim.DensityHist: Stores subcube densities with logarithmic indices.

HPSAnalysis.computeDihedralAngles — Method

computeDihedralAngles(Sim::SimData{R,I}) where {R<:Real, I<:Integer}

Computes the dihedral angles (torsion angles).

Arguments:

Sim::SimData{R,I}: A simulation data structure containing the Simulation information.

Creates:

Sim.TorsionAngles with computed dihedral angles.

HPSAnalysis.computeDihedralHist — Method

computeDihedralHist(Sim::SimData{R,I}; N = 500) where {R<:Real, I<:Integer}

Computes a histogram of dihedral angles.

Arguments:

Sim::SimData{R,I}: A simulation data structure containing the Simulation information.

Creates:

Sim.TorsionHist with the invert of NDihedrals.

HPSAnalysis.computeIntraChainContactMatrix — Method

computeIntraChainContactMatrix(Sim::SimData{R,I})

Computes the intra chain contact map by averaging over all pair distances first and taking the logarithm thereof.

Arguments:

Sim::SimData{R,I}: A simulation data structure containing the Simulation information.

Create:

Sim.IntraChainContactMatrix

HPSAnalysis.computeIntraChainScalingNaiv — Method

computeIntraChainScalingNaiv(Sim::SimData{R,I})

Computes the square intra chain distance |ri-rj| for all combinations and averages over the trajectory.

Arguments:

Sim::SimData{R,I}: A simulation data structure containing the Simulation information.

Create:

Sim.IntraChainScalingNaiv

HPSAnalysis.computeIntraChainScalingSlidingWindow — Method

computeIntraChainScalingSlidingWindow(Sim::SimData{R,I})

Computes the square intra chain distance |ri-rj|^2 and averages in a sliding window approach.

Arguments:

Sim::SimData{R,I}: A simulation data structure containing the Simulation information.

Create:

Sim.IntraChainScalingSlidingWindow

HPSAnalysis.computeSlabDensities — Method

computeSlabHistogram(Sim::SimData; Width=25, MaxVal=0.9, Surface_fac=0.8)

Computes average density within dense phase and dilute phase as well as the indices below/above which Sim.SlabHistogramSeries is in dense/dilute phase.

A mirror symmetric density around zero is computed from which dense phase approximation ρ_app is defined as the mean value of the first Width steps. The dense phase boundary is Surface_fac times the distance r_dense at which the density drops below MaxVal times ρ_app. Similarly the dilute phase boundary is the distance at which the density drops below 1-MaxVal times ρ_app plus r_dense times (1-Surface_fac).

HPSAnalysis.computeSlabHistogram — Method

computeSlabHistogram(Sim::SimData{R,I}; Use_Alpha=false, Use_Types=false) where {R<:Real, I<:Integer}

Computes centered slab histograms along Sim.SlabAxis.

Centered slab histograms are computed for all amino acids, only the positive and only the negatives at all times. Use_Alpha enables the computation of alpha helices and their own slab histogram. Amino acids specific histograms are enabled through Use_Types.

Results are not return but stored in Sim.SlabHistogramSeries as an Offset array where the first dimension ranges from -boxwidth/Sim.Resolution:Sim.Resolution:boxwidth/Sim.Resolution. Default Sim.Resolution is set to $1~\AA$. The second index are the steps at which clusters and slab histogram were computed according to Sim.ClusterRange. The third index are the amino acids which have been used the order in which they are mentioned above. If used, amino acid specific histograms have the index 4+Sim.IDs.

Arguments:

Sim::SimData{R,I}: A simulation data structure containing the Simulation information.

Creat:

Sim.SlabHistogramSeries: Stores mass densities across slabs for different atom types.

HPSAnalysis.getLargestClusterIDs — Method

getLargestClusterIDs(Sim::SimData{T,I}) where {I<:Integer, T<:Real}

Computes IDs for the largest clusters.

Warning is emitted if less than 80% of the proteins are within the largest cluster. The cluster definition might not be reasonable when being used to center slab simulations or other tasks.

Arguments:

Sim::SimData{R,I}: A simulation data structure containing the Simulation information.

Returns:

LClustID::Vector{Int}: List of IDs for largest clusters.

HPSAnalysis.plotAvgSlabDensity — Method

plotAvgSlabDensity(Sim::SimData; Windowlength=100)

Plots the average slab histogram of the frames in range Sim.NSteps-Windowlength:Sim.NSteps. Result is empty if the Windowlength is larger then the step size at which histgrams here computed.

HPSAnalysis.plotAvgSlabDensityEvolution — Method

plotAvgSlabDensityEvolution(Sim::SimData{R,I}; Windowlength=100) where {R<:Real, I<:Integer}

Plot of the average density in the simulation box.

Plots the average slab histogram in range n*Windowlength:(n+1)*Windowlength starting at Sim.EquilibrationTime. Some slices will be empty if the Windowlength is larger then the step size at which histgrams were computed.

Arguments:

Sim::SimData{R,I}: A simulation data structure containing the Simulation information.

Windowlength::Int: How much steps are processed.

Create:

A histogram of the average density.

HPSAnalysis.plotBondAngleHist — Method

plotBondAngleHist(Sim::SimData{R,I})

Plot of the histogram of bond angles.

Arguments:

Sim::SimData{R,I}: A simulation data structure containing the Simulation information.

Create:

A histogram of bond angles.

HPSAnalysis.plotDensityHistogram — Method

plotDensityHistogram(Sim::SimData{R,I}) where {R<:Real, I<:Integer}

Plots the logarithmic density histogram computed by computeDensityHistogram.

Arguments:

Sim::SimData{R,I}: A simulation data structure containing the Simulation information.

Create:

A time averaged histogram of the average density per subcube of the simulation box.

HPSAnalysis.plotDihedralAngleHist — Method

plotDihedralAngleHist(Sim::SimData{R,I})

Plot of the befor computed histogram of dihedral angles.

Arguments:

Sim::SimData{R,I}: A simulation data structure containing the Simulation information.

Create:

A histogram of dihedral angles.

HPSAnalysis.plotIntraChainScaling — Method

plotIntraChainScaling(Sim::SimData{R,I}) where {R<:Real, I<:Integer}

Plots the square intra chain distance |ri-rj| on logarithmic scales as computed by computeIntraChainScalingNaiv and computeIntraChainScalingSlidingWindow.

Arguments:

Sim::SimData{R,I}: A simulation data structure containing the Simulation information.

HPSAnalysis.Setup.BuildENMModel — Method

BuildENMModel(Sim::HPSAnalysis.SimData{T,I}, DomainDict, Proteins, Sequences, ProteinJSON) where {T<:Real, I<:Integer}

Calculate the Indices, that are nessesary to creat a start file im HOOMD.

Arguments

Sim::HPSAnalysis.SimData{T,I}: The simulation datas.

DomainDict: The Domains in which the ENM is active.

Proteins: List of Protein Names.

Sequences: The Sequences of the Proteins.

ProteinJSON: AlphaFold data of the Proteins.

Return:

Number of bonds

Vector of bond type names

ID connecting bond tuple to bond type

Vector of tuples defining all bonds

Dict{String, Dict{Symbol, Float64}} which defines the bonds as used in HOOMD.

HPSAnalysis.Setup.CalvadosSetup — Method

CalvadosSetup(Sequences,AtomTypes,pH)

Define escential Parameters for the Simulation based on the Simulation Type.

Arguments

Sequences::Array{String}: List of sequences of Proteins.

SimulationType::String: Type of Simulation (e.g.: "Calvados2").

pH::Float: pH-value of the system.

Return: A tuple containing:

AtomTypes::Set{Char}: Set of unique amino acid types in the provided sequences.

LongAtomTypes::Set{Char}: Set of amino acid types where the first and last amino acid in a sequence are treated as distinct types.

AaToId::Dict{Char, Int32}: Dictionary mapping each amino acid type to a unique ID number.

IdToAa::Dict{Int32, Char}: Dictionary mapping each ID number to its corresponding amino acid type.

ResToLongAtomType::Dict{Tuple{Char, Bool}, Char}: Dictionary mapping standard amino acids to their modified forms when at the beginning or end of a sequence.

LongAtomTypesToRes::Dict{Char, Tuple{Char, Bool}}: Reverse mapping of ResToLongAtomType.

OneToCharge::Dict{Char, Float}: Dictionary containing the charge values of amino acids, modified based on simulation type.

OneToMass::Dict{Char, Float}: Dictionary containing the mass values of amino acids.

OneToSigma::Dict{Char, Float}: Dictionary containing the sigma values of amino acids.

OneToLambda::Dict{Char, Float}: Dictionary containing the lambda values of amino acids.

OneToHPSDihedral0110::Dict{Char, Any}: Dictionary containing dihedral parameters for a specific configuration.

OneToHPSDihedral1001::Dict{Char, Any}: Dictionary containing dihedral parameters for another configuration.

Notes:

If SimulationType is "Calvados2", the first and last amino acids in each sequence are assigned different types to account for altered mass due to peptide bonding. Also the charge of histidine ('H') is adjusted based on the provided pH value using the formula: 1/(1+10^(pH-6)).

HPSAnalysis.Setup.CombineBackboneAndENM — Method

CombineBackboneAndENM(Proteins, Sequences, HOOMD_Indices, UnfoldedRegions, BackboneCorrectionDict)

Combines the bond length potential of backbone and ENM.

Arguments

Proteins: List of Protein Names.

Sequences: The Sequences of the Proteins.

HOOMD_Indices: Tuple containing the info of ENM as defined by this functions return.

UnfoldedRegions: Dictionary defining the unfolded regions.

BackboneCorrectionDict::Dictionary with atoms, bond length and id which will be connected via backbone.

Return:

Number of bonds

Vector of bond type names

ID connecting bond tuple to bond type

Vector of tuples defining all bonds

Dict{String, Dict{Symbol, Float64}} which defines the bonds as used in HOOMD.

HPSAnalysis.Setup.ComputeHOOMD_ENMIndices — Method

ComputeHOOMD_ENMIndices(ConstraintDict, BackboneCorrectionDict, Sequences, Proteins)

Calculate the Indices, that are nessesary to creat a start file im HOOMD, .

Arguments

`ConstraintDict::Dictionary with atoms, bond length and id which will be connected via ENM.

BackboneCorrectionDict::Dictionary with atoms, bond length and id which will be connected via backbone.

Sequences: The Sequences of the Proteins.

Proteins: List of Protein Names.

Return:

Number of bonds

Vector of bond type names

ID connecting bond tuple to bond type

Vector of tuples defining all bonds

Dict{String, Dict{Symbol, Float64}} which defines the bonds as used in HOOMD.

HPSAnalysis.Setup.DetermineAtomTypes — Method

DetermineAtomTypes(Sequences, SimulationType, pH; OneToChargeDef=BioData.OneToHPSCharge, OneToLambdaDef=BioData.OneToCalvados2Lambda, OneToSigmaDef=BioData.OneToHPSCalvadosSigma)

Defines essential Parameters for the simulation based on the type of the simulation.

Arguments

Sequences::Array{String}: List of sequences of proteins.

SimulationType::String: Type of simulation (e.g.: "Calvados2").

pH::Float: pH-value of the system.

Optional Arguments:

OneToChargeDef::Dict(): Dictionary defining the charge for the Aminoacids.

OneToLambdaDef::Dict(): Dictionary defining the Lambda for the Aminoacids.

OneToSigmaDef::Dict(): Dictionary defining the Sigma for the Aminoacids.

Return: A tuple containing:

AtomTypes::Set{Char}: Set of unique amino acid types in the provided sequences.

LongAtomTypes::Set{Char}: Set of amino acid types where the first and last amino acid in a sequence are treated as distinct types.

AaToId::Dict{Char, Int32}: Dictionary mapping each amino acid type to a unique ID number.

IdToAa::Dict{Int32, Char}: Dictionary mapping each ID number to its corresponding amino acid type.

ResToLongAtomType::Dict{Tuple{Char, Bool}, Char}: Dictionary mapping standard amino acids to their modified forms when at the beginning or end of a sequence.

LongAtomTypesToRes::Dict{Char, Tuple{Char, Bool}}: Reverse mapping of ResToLongAtomType.

OneToCharge::Dict{Char, Float}: Dictionary containing the charge values of amino acids, modified based on simulation type.

OneToMass::Dict{Char, Float}: Dictionary containing the mass values of amino acids.

OneToSigma::Dict{Char, Float}: Dictionary containing the sigma values of amino acids.

OneToLambda::Dict{Char, Float}: Dictionary containing the lambda values of amino acids.

OneToHPSDihedral0110::Dict{Char, Any}: Dictionary containing dihedral parameters for a specific configuration.

OneToHPSDihedral1001::Dict{Char, Any}: Dictionary containing dihedral parameters for another configuration.

Notes:

If SimulationType is "Calvados2", the first and last amino acids in each sequence are assigned different types to account for altered mass due to peptide bonding. Also the charge of histidine ('H') is adjusted based on the provided pH value using the formula: 1/(1+10^(pH-6)).

If SimulationType is "HPS-Alpha", predefined values for charge, lambda, and sigma are used.

If an unknown SimulationType is provided, the function falls back on the user-supplied dictionaries for charge, lambda, and sigma values.

HPSAnalysis.Setup.DetermineCalvados3ENMfromAlphaFold — Method

DetermineCalvados3ENMfromAlphaFold(BasePath, DomainDict, Proteins, ProteinJSON; BBProtein="CA", rcut = 9.0, plDDTcut=90.0, pae_cut=1.85)

Return a dictionary of atoms and there distances that are nessesary for the Elastic Network Model, and a dictionary for the correction of the backbone bonds in the ENM domain.

Arguments

BasePath: The base path of the simulation setup procedure.

DomainDict:: The Domains in which the ENM is active.

Proteins: List of Protein Names.

ProteinJSON: Dictionary of AlphaFold data of the Proteins in JSON format.

BBProtein: The atom from which the AlphaFold datas are set for the aminoacid.

rcut: Cut of length for the ENM in Angstroem.

plDDTcut: Cut of plDDT parameter of AlphaFold reference for the ENM.

pae_cut: Cut of pae parameter of AlphaFold reference for the ENM.

Return:

ConstraintDict: Dictionary that maps Protein names to a Vector containing Tuples of Indices of i,j and distance r which define additional bonds necessary for ENM.

BackboneCorrectionDict: Dictionary that maps Protein names to a Vector containing Tuples of Indices of i,j and distance r which define the backbone bonds that may have different lengths in ENM regions.

HPSAnalysis.Setup.DetermineYukawaInteractions — Method

DetermineYukawaInteractions(;SimulationType="", Temperature=300, SaltConcentration=-1)

Calculates the constants for the Yukawa potencial for given temperature and salt concentration.

Arguments

SimulationType::String: Type of Simulation (e.g.: "Calvados2").

Optional Arguments:

Temperature::Float: Temperature of the Simulation.

SaltConcentration::Float: Salt concentration of the Simulation.

Return:

ϵ_r::Float: Temperature-dependent dielectric constant.

κ::Float: Invers of the Debye length.

HPSAnalysis.Setup.GenerateUnfoldedRegions — Method

GenerateUnfoldedRegions(Proteins, DomainDict, Sequences)

Generates a dictionary that defines the unfolded regions of proteins based on the definition of folded regions.

Arguments

Proteins: List of Protein Names.

DomainDict:: The Domains in which the ENM is active.

Sequences: The Sequences of the Proteins.

Return:

UnfoldedDict:: Dict{String, Vector{Tuple{Int64, Int64}}}(): Maps Protein names onto vectors of tuples that define the regions of unfolded regions.

HPSAnalysis.Setup.WriteDictionaries — Method

WriteDictionaries(filename::String, ToCharge, ToID, ToMass, ToDiameter, ToLambda)

Write a data file with all Dictionaries that are used in the Simulation.

Arguments

filename::String: Path where the data will be saved.

ToCharge::Dict(): Dictionary defining the charge for the Aminoacids.

ToID::Dict(): Dictionary mapping each amino acid type to a unique ID number.

ToMass::Dict(): Dictionary defining the mass for the Aminoacids.

ToDiameter::Dict(): Dictionary defining the diameter for the Aminoacids.

ToLambda::Dict(): Dictionary defining the Lambda for the Aminoacids.

Creat:

Write a file with all Dictionaries.

HPSAnalysis.Setup.WriteENM_HOOMD_Indices — Method

WriteENM_HOOMD_Indices(filename::String, ENM)

Write a data file with all necessary datas for the ENM in HOOMD.

Arguments

filename::String: Path where the data will be saved.

ENM::Tuple: The necessary datas.

Creat:

Write a file with all datas for the ENM.

HPSAnalysis.Setup.WriteHOOMDParticlesInput — Method

WriteHOOMDParticlesInput(filename::String, pos::Array{R}, ToCharge, ToID, Sequences, ToMass, ToDiameter, image) where {R<:Real}

Write a data file that contains position, charge, mass and diameter for the the amino acids.

Arguments

filename::String: Path where the data will be saved.

pos::Array: List of positions for each amino acid.

Sequences::Array{String}: List of protein sequences.

ToMass::Dict(): Dictionary defining the mass for the amino acids.

ToDiameter::Dict(): Dictionary defining the diameter for the amino acids.

image::Array{<Integer}: Number of the periodic image of each amino acid.

Creat:

Write a file with the escential datas of each Aminoacid.

HPSAnalysis.Setup.WriteHOOMDSequences — Method

WriteHOOMDSequences(filename::String, Sequences)

Write a data file with the sequence of each protein.

Arguments

filename::String: Path where the data will be saved.

Sequences::Array{String}: List of protein sequences.

Creat:

Write a file with all Sequences.

HPSAnalysis.Setup.WriteParams — Method

WriteParams(filename, SimName, Temp, NSteps, NOut, Timestep, Box, Seed; Minimise=true, TrajectoryName="traj.gsd", UseAngles=true, UseCharge=true, Alt_GSD_Start="-", Create_Start_Config=false, ϵ_r=1.73136, κ=1.0, Device="GPU", yk_cut=4.0, ah_cut=2.0, ionic=0.1, pH=7.0, SimType="Calvados2",domain=Array([[0,0]]))

Write a data file that contains all Parameters of the Simulation.

Arguments

filename::String: Path where the data will be saved.

SimName::String: Name of the Simulation.

Temp::Float: Temperature of the Simulation.

NSteps::Float: The Number of steps for the Simulation.

NOut::Integer: The Number of steps to determine datas.

Timestep::Float: Time step of the simulation.

Box::Array: Size of the Simulation Box.

Seed::Float: Seed for the pseudorandom number generator.

Minimise::Boolean: Sould the energy be minimize befor the Simulation.

TrajectoryName::String: Data file for the trajectorys.

UseAngles::Boolean: (De-)Activate angle and dihedral potentials (default=off).

UseCharge::Boolean: (De-)Activate yukawa potentials (default=on).

Alt_GSD_Start::String: Name of alternative GSD start file.

Create_Start_Config::Boolean: Set to true, if the python scripts should create a start configuration (numerically unstable/badly equilibrated).

ϵ_r::Float: Dielectric constant.

κ::Float: Debye length of electrostatic potential. .

Device::String: Device on which simulation is running (default=GPU).

yk_cut::Float: Cutoff for the Yukawa potential.

ah_cut::Float: Cutoff for the Ashbaugh potential.

ionic::Float: Ionic strength used for Yukawa potential.

pH::Float: PH value of the simulation.

SimType::String: Type of the simulation (like Calvados2, Calvados3).

domain::Array: Area in with the ENM is active (only for Calvados3).

Creat:

Write a file with all parameters of the simulation that are given from the arguments.

HPSAnalysis.Setup.writeGSDStartFile — Method

writeGSDStartFile(FileName::String, NAtoms::I, NBonds::I, NAngles::I, NDihedrals::I,Box::Vector{R}, Positions::Array{R}, AaToId::Dict{Char, <:Integer},Sequences,  InputImage::Array{I2}, InputMasses::Array{<:Real}, InputCharges::Array{R}, DihedralMap::Dict, DihedralList::Matrix{<:Integer}, AaToSigma::Dict{Char, <:Real}, UseAngles::Bool, SimulationType::String, Domains, ENM) where {I<:Integer, R<:Real, I2<:Integer}

A GSD data file is written, that include the parameters for the Simulation witch are given as Arguments.

Arguments

FileName::String: Name of the output GSD file.

NAtoms::Int: The total number of amino acids (atoms) in the system.

NBonds::Int: The number of bonds between amino acids (NAtoms - 1).

NAngles::Int: The number of angles formed between amino acids (NAtoms - 2).

NDihedrals::Int: The number of dihedral angles between amino acids (NAtoms - 3).

Box::Vector{Float}: 3-element vector specifying the dimensions of the simulation box

Positions::Array{Float}: Array of the atomic coordinates of the Aminoacids.

AaToId::Dict{Char, Int}: Dictionary mapping each amino acid type to a unique ID number.

Sequences::Array{String}: A list of protein sequences, where each sequence is represented as a string of amino acids.

InputImage::Array{Int}: An array used to determine periodic boundary conditions and correct atomic positions.

InputMasses::Array{Float}: A 1D array specifying the mass of each amino acid.

InputCharges::Array{Float}: A 1D array specifying the electric charge of each amino acid.

DihedralMap::Dict: A dictionary mapping unique dihedral angle definitions (four atom indices) to dihedral types.

DihedralList::Matrix{Int}: A matrix where each row defines a specific dihedral angle using atom indices.

AaToSigma::Dict{Char, <:Real}: A dictionary mapping amino acid types to their Lennard-Jones σ-parameter (used in force field calculations).

UseAngles::Bool: If true, angle and dihedral interactions are included in the GSD file.

SimulationType::String: Type of simulation.

Domains::List(Int): Domains in which the ENM is active.

ENM::Tuple: Data that are nessesary for ENM (bond name, id, group).

Creates:

Writes the GSD data files.

HPSAnalysis.Setup.writeHOOMD — Method

writeHOOMD(Sequences,pos,image,OneToCharge,AaToId,OneToMass,OneToSigma,OneToLambda,AlphaAddition,dihedral_long_map,dihedral_eps,SimulationType,Temperature,SaltConcentration,BoxSize,StartFileName,NSteps,WriteOutFreq,Device,yk_cut,ah_cut,pH,domain,NAtoms,NBonds,NAngles,NDihedrals,dihedral_short_map,dihedral_list, ENM)

Writes the datas for the simulation in different files with HOOMD.

Arguments

from writeStartConfiguration()

Creates:

Writes the datas for the simulation.

HPSAnalysis.Setup.writeHPSLammps — Method

writeHPSLammps(fileName,Sequences,AtomTypes,LongAtomTypes,LongAtomTypesToRes,InitStyle,ChargeTemperSteps,ChargeTemperSwapSteps,pos,image,OneToCharge,AaToId,OneToMass,OneToSigma,OneToLambda,AlphaAddition,dihedral_long_map,dihedral_eps,SimulationType,Temperature,SaltConcentration,BoxSize,StartFileName,NSteps,WriteOutFreq,pH,NAtoms,NBonds,NAngles,NDihedrals,Info,NAtomTypes)

Writes the datas for the simulation in different files.

Arguments

from writeStartConfiguration()

Creates:

Writes the datas for the simulation.

HPSAnalysis.Setup.writeStartConfiguration — Function

writeStartConfiguration(fileName, StartFileName, Info, Sequences, BoxSize,NSteps=100_000_000; SimulationType="Calvados2", Temperature=300,MixingRule="1-1001-1", Pos =zeros(Float32, 0),InitStyle="Slab", SaltConcentration=0.15, pH=6.0, ChargeTemperSteps=[], ChargeTemperSwapSteps=100_000, HOOMD=false, OneToChargeDef=BioData.OneToHPSCharge, OneToLambdaDef=BioData.OneToCalvados2Lambda, OneToSigmaDef=BioData.OneToHPSCalvadosSigma,WriteOutFreq=100_000, Device="GPU", yk_cut=40.0, ah_cut=20.0)

Writes the start configuration for a molecular dynamics simulation.

Arguments

fileName::String: Name of the output file.

StartFileName::String: Name of the initial configuration file.

Info::String: Additional information about the simulation.

Sequences::Vector{String}: List of amino acid sequences corresponding to the proteins.

BoxSize::Vector{Float}: A vector of minmal/maximal box dimensions in each axis. ([xmin, xmax, ymin, ymax, zmin, zmax]).

NSteps::Int: Number of simulation steps (default: 100,000,000).

SimulationType::String: IDP model used for simulations.(default: "Calvados2").

Temperature::Float: Temperature in Kelvin (default: 300).

MixingRule::String: Mixing rule for optional dihedral potential for Calvados models.

Pos::Vector{Float}: Initial positions. (default: empty array).

InitStyle::String: Initialization style, e.g., "Slab" or "Pos".

SaltConcentration::Float: Salt concentration in M (default: 0.15).

pH::Float: pH level (default: 6).

ChargeTemperSteps: List of charge tempering steps.

ChargeTemperSwapSteps::Int: Swap steps for charge tempering.

HOOMD::Boolean: Boolean to enable HOOMD compatibility.

OneToChargeDef::Dict: Dictionary defining the amount of charge for each one letter atom type.

OneToSigmaDef::Dict: Dictionary defining the sigma parameter in LJ potentials for each one letter atom type.

OneToLambdaDef::Dict: Dictionary defining the lambda paramter in Ashbaugh-Hatch potentials for each one letter atom type.

WriteOutFreq::Int: Frequency of writing output (default: 100,000).

Device::String: Computational device, "CPU"/"GPU" (default: "GPU").

yk_cut::Float: Cutoff distances for yukawa potential.

ah_cut::Float: Cutoff distances for ashbaugh hatch potential.

domain::List(Int): Domains in which the ENM is active.

ENM::Tuple: Data that are nessesary for ENM (bond name, id, group).

Creates:

Writes data files with the start configuration.

HPSAnalysis.Polyply.ConvertGroToPDB — Method

ConvertGroToPDB(Path, Filename)

A gro data file is converted to a pdb data file.

Arguments

Path::String: Path to the gro file.

Filename::String: Name of the gro file.

Creat:

A pdb data file.

HPSAnalysis.Polyply.GenerateCoordinates — Method

GenerateCoordinates(SimulationPath, SimulationName, Box, TopologyFile)

Generate a gro file from the TopologyFile which is used for the start coordinates.

Arguments

SimulationPath::String: Path where to save the gro data file.

SimulationName::String: Name of the simulation (save name).

Box::Arry(Float): Diameters of the simulationbox.

TopologyFile::String: Topology data file.

Creat:

A gro data file.

HPSAnalysis.Polyply.GenerateENM_ITPFilesOfSequence — Method

GenerateENM_ITPFilesOfSequence(BasePath::String, Names, Domains::Dict{String,Vector{Tuple{I,I}}})

Generate pdb and itp data files for the ENM with martinize2 (only Calvados3).

Arguments

BasePath::String: Data path.

Names::List(String): List of protein names.

Domains::Dict{String,Vector{Tuple{I,I}}}: Dictionary with domains for each protein.

Creat:

Data files for the ENM.

HPSAnalysis.Polyply.GenerateITPFilesOfSequence — Method

GenerateITPFilesOfSequence(Names, Sequences, OutputPath)

Generate a itp file for one protein with the given sequence, with datas like atoms, bonds, angels.

Arguments

Names::Array(String): Names of the proteins.

Sequences::Array(String): Sequences of the proteins.

OutputPath::String: Path for the itp file.

Creat:

A itp file.

HPSAnalysis.Polyply.GenerateSlabTopologyFile — Method

GenerateSlabTopologyFile(Filename, ITPPath, Names, SimulationName)

Generate a Topology file which include the number and name of the proteins that are simulated.

Arguments

Filename::String: Path for the topology file.

ITPPath::String: Pathe of the itp file.

Names:Array(String):: List of protein names.

SimulationName::String: Name of the simulation.

Creat:

A topology data file.

HPSAnalysis.Polyply.readPDB — Method

readPDB(Filename, x,y,z)

Read the coordinates of the atoms from a pdb file.

Arguments

Filename::: Path of the pdb file.

x::Array(Float): List of x-coordinates.

y::Array(Float): List of y-coordinates.

z::Array(Float): List of z-coordinates.

HPSAnalysis.Polyply.readSimpleGRO — Method

readSimpleGRO(Filename, x,y,z)

Read the coordinates of the atoms from a gro file.

Arguments

Filename::: Path of the pdb file.

x::Array(Float): List of x-coordinates.

y::Array(Float): List of y-coordinates.

z::Array(Float): List of z-coordinates.